pyproject.toml
[build-system]
requires = ["flit_core >=3.11,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "checkmate"
version = "0.1.0"
authors = [
  { name = "Checkmate Team" },
]
license = "Apache-2.0"
license-files = ["LICENSE"]
description = "Model-agnostic LLM vulnerability scanner"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
  "base2048>=0.1.3",
  "transformers>=4.51.3,<4.57.0",
  "datasets>=3.0.0,<4.0",
  "colorama>=0.4.3",
  "tqdm>=4.67.1",
  "cohere>=5.16.0",
  "openai>=1.45.0,<2",
  "replicate>=0.8.3",
  "google-api-python-client>=2.0",
  "backoff>=2.1.1",
  "rapidfuzz>=3.0.0",
  "jinja2>=3.1.6",
  "nltk>=3.9.1",
  "accelerate>=0.23.0",
  "avidtools==0.1.2",
  "stdlibs>=2022.10.9",
  "langchain>=0.3.25",
  "nemollm>=0.3.0",
  "cmd2==2.4.3",
  "torch>=2.6.0",
  "sentencepiece>=0.1.99",
  "markdown>=3.4.3",
  "numpy>=2.0.0",
  "zalgolib>=0.2.2",
  "ecoji>=0.1.1",
  "deepl==1.17.0",
  "fschat>=0.2.36",
  "litellm>=1.68.1",
  "jsonpath-ng>=1.6.1",
  "huggingface_hub>=0.21.0",
  'python-magic-bin>=0.4.14; sys_platform == "win32"',
  'python-magic>=0.4.21; sys_platform != "win32"',
  "lorem==0.1.1",
  "xdg-base-dirs>=6.0.1",
  "wn==0.9.5",
  "ollama>=0.4.7",
  "nvidia-riva-client==2.16.0",
  "google-cloud-translate>=2.0.4",
  "grpcio-tools>=1.71.0",
  "langdetect==1.0.9",
  "tiktoken>=0.7.0",
  "mistralai==1.5.2",
  "pillow>=10.4.0",
  "ftfy>=6.3.1",
  "pydantic>=2.0.0",
  "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-mock>=3.14.0",
  "requests-mock==1.12.1",
  "respx>=0.21.1",
  "pytest-cov>=5.0.0",
  "pytest_httpserver>=1.1.0",
  "langcodes>=3.4.0",
  "black==24.4.2",
  "ruff>=0.1.0",
  "mypy>=1.0.0",
  "isort>=5.0.0",
  "pre-commit>=3.0.0",
  "bandit>=1.7.0",
  "pip-audit>=2.6.0",
]
tests = [
  "pytest>=8.0",
  "pytest-mock>=3.14.0",
  "requests-mock==1.12.1",
  "respx>=0.21.1",
  "pytest-cov>=5.0.0",
  "pytest_httpserver>=1.1.0",
  "langcodes>=3.4.0",
]
lint = [
  "black==24.4.2",
  "ruff>=0.1.0",
  "mypy>=1.0.0",
  "isort>=5.0.0",
]
calibration = [
  "scipy>=1.14.0",
]
audio = [
  "soundfile>=0.13.1",
  "librosa>=0.10.2"
]
dra = [
  "detoxify>=0.5.0"
]

[project.urls]
"Homepage" = "https://github.com/checkmate-ai/checkmate"
"Bug Tracker" = "https://github.com/checkmate-ai/checkmate/issues"

[project.scripts]
checkmate = "checkmate.cli:main"

[tool.flit.module]
name = "checkmate"

[tool.flit.sdist]
include = [
  "*.md",
  "*.txt",
  "checkmate/analyze/templates/*",
  "checkmate/configs/*",
  "checkmate/resources/*.json",
  "checkmate/resources/*.yaml",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore",
    "default:::checkmate",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
    "-ra",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["./checkmate"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]